move var tests
